Translate default messages
