+++ serverthrall/localization/translations.py
@@ -1,0 +1,25 @@
+# flake8: noqa
+
+language_map = {
+    'en': {
+        'minute': 'minute',
+        'minutes': 'minutes',
+        'never': 'never',
+        'ServerRestarter.warning': 'The server is being restarted in $timeleft $timeunit.',
+        'ServerRestarter.restart': 'The server is being restarted now.',
+        'ServerUpdater.warning': 'There is a new update available. The Server is restarting in $timeleft $timeunit to install the update.',
+        'ServerUpdater.restart': 'The server is restarting to install an update!',
+    },
+    'pl': {
+        'minute': 'minuta',
+        'minutes': 'minut',
+        'never': 'nigdy'
+    },
+    'es': {
+        'minute': 'minuto',
+        'minutes': 'minutos',
+        'never': 'nunca',
+        'ServerUpdater.warning': 'Hay una nueva actualización disponible. El servidor va a reiniciarse en $timeleft $timeunit para instalr la actualización',
+        'ServerUpdater.restart': 'El servidor se está reiniciando para instalar una actualización!',
+    }
+}
